FEAT: Remove project prefix from file names
All files in project "projectname" have file names like
"projects/projectname/foo.vcf.gz", which is verbose and redundant.

Change the representation of project files so that it doesn't include
the prefix "projects/projectname/". This means removing the prefix from
the deletion and project files update code, and adding it back in when
creating the presigned url.