FEAT: Update project files on s3 changes
When files are added to s3, the project item in dynamodb isn't updated.
This means it can become out of sync if files are added through the
console rather than the frontend. In addition, it is desired that the
total number of samples across all vcf/bcf files in a project should be
counted.

Create the updateFiles lambda function to keep track of the files in
each project, as well as the total samples across vcf/bcf files in each
project. This function will store the sample numbers for each file, as
well as update the projects dynamodb table. It also offers refresh
functionality, which is currently only accessible through direct
invocation, and should only be used if a project is stuck in an
inconsistent state. In this case the input json should contain a
"project" key with the value being the name of the project to refresh.