REFACTOR: Don't edit files in project creation
The project create admin lambda function reads the files argument and
sets it accordingly. However the file setting is now handled by the
updateFiles lambda function, triggered directly by changes in s3.

Remove references to files in the project creation code, to avoid
stepping on the toes of the s3 integration.