FEAT: Filter all sBeacon queries by projects
Users should be able to query sBeacon in a way that specifies one or
more projects, and only return values for those projects. In addition,
users should not be able to query the details of any projects to which
they have no access.

Take advantage of the common logic when querying athena to hook into all
queries. Add WHERE clauses to queries over any of the supported tables
(datesets, individuals, biosamples, runs. and analyses) that specify
that the project must be in the intersection of requested projects and
projects to which the user has access. To this end, add the user id
(sub) to the request object, and pass this sub and the requested
projects into the run_custom_query function, which is used by all the
sBeacon querying functions.

This impacts 81 different athena queries, so there may be bugs,
especially in the parsing logic. The logic is not perfect, and will be
tripped up by particularly complex queries, but has been designed so
that it should crash in such cases rather than opening up a security
hole. Of particular note is the /datasets/{id}/filterering_terms
endpoint, which is quite complex and difficult to test because it is
currently non-functional on the frontend. Future queries that may
include a WHERE A OR B clause will also cause the filtering to be
non-functional in the case that B is true. This is because the clause
will be edited such that the clause will become WHERE projectFilter AND
A OR B.