FIX: Encapsulate query "?"s in spaces
Because the add_project_names parsing logic splits by spaces to find
semantic words, if the placeholder for an execution parameter "?" is
immediately preceded or followed by another nonspace character (not
including parentheses, which are stripped), the parser does not
recognise the "?". This means that the associated execution parameter is
not added in the right place and we end up with excess parameters, and a
broken query besides.

Surround each use of the "?" character in query building with space
characters. Future uses of execution parameters that do not do this
should immediately crash under most tests, so these sorts of errors will
not fly under the radar.