FIX: Improve _projectNames tables identification
The full reference of a table (often "database"."tablename") is searched
to see if it mentions any of the tables that need to be filtered by
_projectName. This is done by simply searching to see if any of the
names of the relevant tables are contained with in the reference name.
This can cause problems if the name of one table is a substring of
another table. For example sbeacon_terms_index will trigger because the
table sbeacon_terms needs to be filtered.

Change the method of identification so the table is split on '"' and '.'
chaacters, and then perform an intersection over the two sets of
strings. This could mean that if the database name is also a match of a
table every table in the database will trigger the filter. We could make
an even more robust identification, but I'm not sure how complicated a
reference can become, and we shouldn't be using the same names for
databases and tables anyway.