FEAT: Add _projectname to terms_index table
We filter the terms table by the _projectname column, but not currently
the terms_index table. If we add the _projectname column to the index we
won't need to reference the main table and can perform some SQL pushdown
to narrow the join space when joining across multiple entities using the
index.

Add the _projectname column to the the sbeacon_terms_index table.