FIX: Don't include samples in annotation file
When generating the annotation file to update records whose INFO columns
contain PII, the full file header is loaded in. Because the annotation
file doesn't include any columns beyond INFO, the FORMAT and sample
columns from the main files header conflicts.

Only use the first 8 columns from the final line of the header, namely
those up to the INFO column. Also change the process_records function to
use a copy of the header_lines list, because I don't like the idea of
side effects if we want to use that list later on.