FIX: Find libraries required by bcftools correctly
When searching for libraries required by the bcftools binary for
inclusion in the lambda layer, the bcftools directory is searched
instead of the binary. This has shown no ill effects, perhaps because
the required libraries are already loaded when collecting tabix's
libraries.

Search for bcftools/bcftools rather than just bcftools when collecting
libraries.