FEAT: Add versioning to dataportal bucket
User data should not be able to be easily deleted. If deleted, data
should be accessible again.

Add versioning to the dataportal bucket to meet user requirements.
Because versioning can not be carved out of prefixes in a bucket,
implement as stringent a culling protocol as possible on the staging/
prefix. This is to delete noncurrent versions of objects after 1 day.