FIX: Use directory hash to refresh docker digest
When a code change causes a new hash of the function directory, a new
docker image is created. If this image has the same hash digest as the
previous image (for example a comment in the docker file triggered the
update), terraform will destroy the new image when it destroys the old
image, as it destroys by digest. Alternatively it may silently fail to
update the docker image, as the tag ("latest") and digest will be the
same.

Insert a hash of the directory into the docker image, so its digest will
always be different if the directory changes in any way.