FIX: Report unsorted BAM file error
When a BAM or SAM file is unsorted, indexing as the last step fails.
This is not picked up as a process error as the first sign is that
attempting to print to that process's stdin fails.

Catch the BrokenPipeError when trying to print to the failed process,
and instead check that process's status, likely raising the more usable
ProcessError.