FIX: Don't process uploaded index files at all
When index files are uploaded, they are immediately pushed through to S3
without any deidentification. The idea is that the VCF deidentification
step will then create the correct index file and upload it, overwriting
the original. However this assumes that the VCF event is completed after
the index event, and if the uploaded index is a .tbi file, the new
index, a .csi file, will not overwrite it. This means if a tool prefers
.tbi files to .csi files, it may use the now out-of-date .tbi file. .bai
files are sent through deidentification, but then further processing is
skipped, leaving the files in the staging bucket.

Whenever an index file (.bai, .csi, or .tbi) is uploaded, completely
ignore it and delete it from staging. This will mean the indexes created
when uploading the indexable genomic files will be the only ones present
in the directory, preventing any possible conflicts or out-of-date
indexes.